Update blank.yml
removed 1.4 and 1.5 from the testing pipeline. The issue appears to be that because GeoParams.jl is not yet registered, we need to upload a Manifest.toml file and tell the package manager the GitHub address to download it. This Manifest file was generated using Julia 1.6, which appears incompatible with older versions. The code itself still appears to work on older systems, though