Added functionality to project different datasets onto given profiles profile.
This adds the functionality to project different datasets onto a given profile using existing functions in GMG. Profile data as well as the information on different datasets are given in text files. I also added JLD2 as a dependency to save the different profiles. In principle, Matlab files can also be produced, but this would require a dependency on MATLAB.jl.